+++ files/JSON/thisIsAFile.json
@@ -1,7 +1,7 @@
 {
   "message": "please work :pleading_face:",
   "language": "english, duh",
-  "testing": true,
+  "testing": "true",
 
   "sOFTesting": {
     "comment": "Testing data for steelOrFeathers",
